ref(framestack): break it down to separate files
