+++ static/app/components/profiling/flamegraphOptionsContextMenu.tsx
@@ -13,172 +13,7 @@ import {
 } from 'sentry/utils/profiling/flamegraph/flamegraphStateProvider/flamegraphPreferences';
 import {useFlamegraphPreferences} from 'sentry/utils/profiling/flamegraph/useFlamegraphPreferences';
 import {Rect} from 'sentry/utils/profiling/gl/utils';
-
-export function useContextMenu() {
-  const [open, setOpen] = useState<boolean>(false);
-  const itemProps = useKeyboardNavigation();
-
-  function wrapSetOpen(newOpen: boolean) {
-    if (!newOpen) {
-      itemProps.setTabIndex(null);
-    }
-    setOpen(newOpen);
-  }
-
-  function getMenuProps() {
-    const menuProps = itemProps.getMenuKeyboardEventHandlers();
-
-    return {
-      ...menuProps,
-      onKeyDown: (evt: React.KeyboardEvent) => {
-        if (evt.key === 'Escape') {
-          setOpen(false);
-        }
-        menuProps.onKeyDown(evt);
-      },
-    };
-  }
-
-  function getMenuItemProps() {
-    const menuItemProps = itemProps.getMenuItemKeyboardEventHandlers();
-
-    return {
-      ...menuItemProps,
-      onKeyDown: (evt: React.KeyboardEvent) => {
-        if (evt.key === 'Escape') {
-          setOpen(false);
-        }
-        menuItemProps.onKeyDown(evt);
-      },
-    };
-  }
-
-  return {
-    open,
-    setOpen: wrapSetOpen,
-    menuRef: itemProps.menuRef,
-    getMenuProps,
-    getMenuItemProps,
-  };
-}
-
-function useKeyboardNavigation() {
-  const [menuRef, setMenuRef] = useState<HTMLDivElement | null>(null);
-  const [tabIndex, setTabIndex] = useState<number | null>(null);
-
-  const items: {id: number; node: HTMLElement | null}[] = [];
-
-  useEffect(() => {
-    if (menuRef) {
-      if (tabIndex === null) {
-        menuRef.focus();
-      }
-    }
-  }, [menuRef, tabIndex]);
-
-  useEffect(() => {
-    if (typeof tabIndex !== 'number') {
-      return;
-    }
-    if (items[tabIndex]?.node) {
-      items[tabIndex]?.node?.focus();
-    }
-  }, [tabIndex]);
-
-  function getMenuKeyboardEventHandlers() {
-    return {
-      tabIndex: -1,
-      ref: setMenuRef,
-      onKeyDown: (evt: React.KeyboardEvent) => {
-        if (items.length === 0) {
-          return;
-        }
-
-        if (evt.key === 'Escape') {
-          setTabIndex(null);
-        }
-
-        if (evt.key === 'ArrowDown' || evt.key === 'Tab') {
-          evt.preventDefault();
-
-          if (tabIndex === items.length - 1 || tabIndex === null) {
-            setTabIndex(0);
-          } else {
-            setTabIndex((tabIndex ?? 0) + 1);
-          }
-        }
-
-        if (evt.key === 'ArrowUp' || (evt.key === 'Tab' && evt.shiftKey)) {
-          evt.preventDefault();
-
-          if (tabIndex === 0 || tabIndex === null) {
-            setTabIndex(items.length - 1);
-          } else {
-            setTabIndex((tabIndex ?? 0) - 1);
-          }
-        }
-      },
-    };
-  }
-
-  function getMenuItemKeyboardEventHandlers() {
-    const idx = items.length;
-    items.push({id: idx, node: null});
-
-    return {
-      tabIndex: tabIndex === idx ? 0 : -1,
-      ref: (node: HTMLElement | null) => {
-        if (items[idx]) {
-          items[idx].node = node;
-        }
-      },
-      onMouseEnter: () => {
-        setTabIndex(idx);
-      },
-      onKeyDown: (evt: React.KeyboardEvent) => {
-        if (items.length === 0) {
-          return;
-        }
-
-        if (evt.key === 'Escape') {
-          setTabIndex(null);
-        }
-
-        if (evt.key === 'Enter' || evt.key === ' ') {
-          items?.[idx]?.node?.click?.();
-        }
-
-        if (evt.key === 'ArrowDown' || evt.key === 'Tab') {
-          evt.preventDefault();
-
-          if (tabIndex === items.length || tabIndex === null) {
-            setTabIndex(0);
-          } else {
-            setTabIndex((tabIndex ?? 0) + 1);
-          }
-        }
-
-        if (evt.key === 'ArrowUp' || (evt.key === 'Tab' && evt.shiftKey)) {
-          evt.preventDefault();
-
-          if (tabIndex === 0 || tabIndex === null) {
-            setTabIndex(items.length);
-          } else {
-            setTabIndex((tabIndex ?? 0) - 1);
-          }
-        }
-      },
-    };
-  }
-
-  return {
-    menuRef,
-    getMenuItemKeyboardEventHandlers,
-    getMenuKeyboardEventHandlers,
-    tabIndex,
-    setTabIndex,
-  };
-}
+import {useContextMenu} from 'sentry/utils/profiling/hooks/useContextMenu';
 
 interface MenuProps
   extends React.DetailedHTMLProps<React.HTMLAttributes<HTMLDivElement>, HTMLDivElement> {
